- Fixed file deletion

+++ src/main/java/de/flavormate/ba_entities/file/service/FileService.java
@@ -6,11 +6,15 @@ import de.flavormate.aa_interfaces.services.ICRUDService;
 import de.flavormate.ab_exeptions.exceptions.ConflictException;
 import de.flavormate.ab_exeptions.exceptions.CustomException;
 import de.flavormate.ab_exeptions.exceptions.NotFoundException;
+import de.flavormate.ab_exeptions.exceptions.UnauthorizedException;
 import de.flavormate.ba_entities.account.model.Account;
 import de.flavormate.ba_entities.account.repository.AccountRepository;
+import de.flavormate.ba_entities.file.enums.FileCategory;
 import de.flavormate.ba_entities.file.model.File;
 import de.flavormate.ba_entities.file.repository.FileRepository;
 import de.flavormate.ba_entities.file.wrapper.FileDraft;
+import de.flavormate.ba_entities.recipe.model.Recipe;
+import de.flavormate.ba_entities.recipe.repository.RecipeRepository;
 import org.apache.commons.codec.binary.Base64;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -27,13 +31,15 @@ public class FileService extends BaseService implements ICRUDService<File, FileD
 
 	private final FileRepository repository;
 	private final AccountRepository accountRepository;
+	private final RecipeRepository recipeRepository;
 
 	@Value("${flavorMate.files}")
 	private URL ROOT;
 
-	protected FileService(FileRepository repository, AccountRepository accountRepository) {
+	protected FileService(FileRepository repository, AccountRepository accountRepository, RecipeRepository recipeRepository) {
 		this.repository = repository;
 		this.accountRepository = accountRepository;
+		this.recipeRepository = recipeRepository;
 	}
 
 	@Override
@@ -65,8 +71,27 @@ public class FileService extends BaseService implements ICRUDService<File, FileD
 	@Override
 	public boolean deleteById(Long id) throws CustomException {
 		try {
-			var owner = accountRepository.findByUsername(getPrincipal().getUsername()).orElseThrow(() -> new NotFoundException(Account.class));
-			var file = repository.findById(id, owner.getId()).orElseThrow(() -> new NotFoundException(File.class));
+			var account = accountRepository.findByUsername(getPrincipal().getUsername()).orElseThrow(() -> new NotFoundException(Account.class));
+			var file = repository.findById(id).orElseThrow(() -> new NotFoundException(File.class));
+
+			var authorized = false;
+			if (account.hasRole("ROLE_ADMIN")) {
+				authorized = true;
+			} else {
+				authorized = switch (file.getCategory()) {
+					case FileCategory.ACCOUNT -> file.getOwner().equals(account.getId());
+					case FileCategory.AUTHOR -> false;
+					case FileCategory.RECIPE -> {
+						var recipe = recipeRepository.findById(file.getOwner()).orElseThrow(() -> new NotFoundException(Recipe.class));
+						yield account.getId().equals(recipe.getAuthor().getAccount().getId());
+					}
+				};
+			}
+
+			if (!authorized) {
+				throw new UnauthorizedException(File.class);
+			}
+
 
 			var deleted = Files.deleteIfExists(Paths.get(ROOT.getPath(), file.getPath()));
